+++ homebridge/ring-platform.ts
@@ -27,6 +27,8 @@ import { RingAuth } from '../api/rest-client'
 import { platformName, pluginName } from './plugin-info'
 import { useLogger } from '../api/util'
 import { BaseAccessory } from './base-accessory'
+import { FloodFreezeSensor } from './flood-freeze-sensor'
+import { FreezeSensor } from './freeze-sensor'
 
 const debug = __filename.includes('release-homebridge'),
   unsupportedDeviceTypes: (RingDeviceType | RingCameraKind)[] = [
@@ -46,6 +48,10 @@ function getAccessoryClass(
       return ContactSensor
     case RingDeviceType.MotionSensor:
       return MotionSensor
+    case RingDeviceType.FloodFreezeSensor:
+      return FloodFreezeSensor
+    case RingDeviceType.FreezeSensor:
+      return FreezeSensor
     case RingDeviceType.SecurityPanel:
       return SecurityPanel
     case RingDeviceType.BaseStation:
@@ -180,6 +186,9 @@ export class RingPlatform {
             (this.config.hideUnsupportedServices &&
               unsupportedDeviceTypes.includes(device.deviceType))
           ) {
+            this.log.info(
+              `Hidden accessory ${device.deviceType} ${displayName}`
+            )
             return
           }